requirements.txt
requests
python-dotenv
google-cloud-translate